feat: restructures queries, adds directory to jest testPathIgnorePatterns
